Add GetAllWeights driver

+++ srv/api/weight/model.go
@@ -2,7 +2,7 @@ package weight
 
 // Weight model
 type Weight struct {
-	ID    string `json:"id"`
-	Name  string `json:"name"`
-	Value int    `json:"value"`
+	ID    string  `json:"id"`
+	Name  string  `json:"name"`
+	Value float64 `json:"value"`
 }
